@@ -17,16 +17,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.1.3",
-    "@angular/common": "^4.1.3",
-    "@angular/compiler": "^4.1.3",
-    "@angular/core": "^4.1.3",
-    "@angular/forms": "^4.1.3",
-    "@angular/http": "^4.1.3",
+    "@angular/animations": "^4.2.3",
+    "@angular/common": "^4.2.3",
+    "@angular/compiler": "^4.2.3",
+    "@angular/core": "^4.2.3",
+    "@angular/forms": "^4.2.3",
+    "@angular/http": "^4.2.3",
     "@angular/material": "^2.0.0-beta.6",
-    "@angular/platform-browser": "^4.1.3",
-    "@angular/platform-browser-dynamic": "^4.1.3",
-    "@angular/router": "^4.1.3",
+    "@angular/platform-browser": "^4.2.3",
+    "@angular/platform-browser-dynamic": "^4.2.3",
+    "@angular/router": "^4.2.3",
     "@ngrx/core": "^1.0.0",
     "@ngrx/db": "^2.0.2",
     "@ngrx/effects": "^2.0.3",
@@ -40,21 +40,21 @@
     "hammerjs": "^2.0.8",
     "ng2-md-datatable": "^1.3.15",
     "reselect": "^3.0.1",
-    "rxjs": "^5.3.0",
+    "rxjs": "^5.4.1",
     "ts-helpers": "^1.1.1",
     "zone.js": "^0.8.12"
   },
   "devDependencies": {
-    "@angular/cli": "^1.1.1",
-    "@angular/compiler-cli": "^4.1.3",
-    "@angular/language-service": "^4.2.0-rc.2",
+    "@angular/cli": "^1.1.2",
+    "@angular/compiler-cli": "^4.2.3",
+    "@angular/language-service": "^4.2.3",
     "@angularclass/hmr": "^1.2.2",
     "@types/jasmine": "^2.5.48",
     "@types/node": "^7.0.29",
     "codelyzer": "^3.0.1",
     "eslint": "^4.0.0",
     "gh-pages": "^1.0.0",
-    "jasmine-core": "^2.6.3",
+    "jasmine-core": "^2.6.4",
     "jasmine-spec-reporter": "^4.1.0",
     "karma": "^1.7.0",
     "karma-chrome-launcher": "^2.1.1",
